+++ doc/changelog.rst
@@ -32,6 +32,9 @@ Added
 
 Fixed
 -----
+- Merging of (typically refined) crystal maps, where either a simulation indices
+  array is not present or the array contains more indices per point than scores.
+  (`#335 <https://github.com/pyxem/kikuchipy/pull/335>`_)
 - Bugs in getting plot markers from geometrical EBSD simulation.
   (`#334 <https://github.com/pyxem/kikuchipy/pull/334>`_)
 - Passing a static background pattern to EBSD.remove_static_background() for a